Refactor + add bedrock
