+++ tests/functions/test_bedrock.py
@@ -1,8 +1,10 @@
 import pytest
 
 import pixeltable as pxt
-from tests.conftest import DO_RERUN
-from tests.utils import skip_test_if_no_aws_credentials, skip_test_if_not_installed, stock_price
+
+from ..conftest import DO_RERUN
+from ..utils import skip_test_if_no_aws_credentials, skip_test_if_not_installed, stock_price
+from .tool_utils import run_tool_invocations_test
 
 
 @pytest.mark.remote_api
@@ -43,16 +45,17 @@ class TestBedrock:
     def test_tool_invocations(self, reset_db: None) -> None:
         skip_test_if_not_installed('boto3')
         skip_test_if_no_aws_credentials()
-        from pixeltable.functions.bedrock import converse, invoke_tools
+        from pixeltable.functions import bedrock
 
-        tools = pxt.tools(stock_price)
-        t = pxt.create_table('test_tbl', {'input': pxt.String})
-        messages = [{'role': 'user', 'content': [{'text': t.input}]}]
-        t.add_computed_column(
-            response=converse(messages, model_id='anthropic.claude-3-haiku-20240307-v1:0', tool_config=tools)
-        )
-        t.insert(input='What is the stock price of NVDA today?')
-        t.add_computed_column(tool_calls=invoke_tools(tools, t.response))
+        def make_table(tools: pxt.func.Tools, tool_choice: pxt.func.ToolChoice) -> pxt.Table:
+            t = pxt.create_table('test_tbl', {'prompt': pxt.String})
+            messages = [{'role': 'user', 'content': [{'text': t.prompt}]}]
+            t.add_computed_column(
+                response=bedrock.converse(
+                    messages, model_id='anthropic.claude-3-haiku-20240307-v1:0', tool_config=tools
+                )
+            )
+            t.add_computed_column(tool_calls=bedrock.invoke_tools(tools, t.response))
+            return t
 
-        results = t.collect()[0]
-        assert results['tool_calls']['stock_price'] == [131.17]
+        run_tool_invocations_test(make_table, test_multiple_tool_use=False)
